refactor: routes & theme  constants
